remote-access-app/requirements.txt
# Dependências do Remote Access App
# Instale com: pip install -r requirements.txt

# ========================
# NÚCLEO
# ========================
cryptography>=41.0.0          # Criptografia AES-256-GCM
Pillow>=10.0.0                # Processamento de imagens

# ========================
# CAPTURA DE TELA
# ========================
mss>=9.0.1                    # Screenshot eficiente em Windows
numpy>=1.24.0                 # Manipulação de arrays

# ========================
# INTERFACE GRÁFICA (Futuro - Opcional para MVP)
# ========================
# PyQt5>=5.15.0               # Interface gráfica (descomente na v2.0)

# ========================
# DESENVOLVIMENTO E TESTES
# ========================
pytest>=7.4.0                 # Framework de testes
pytest-asyncio>=0.21.0        # Suporte a testes assíncronos
black>=23.0.0                 # Formatação de código
flake8>=6.0.0                 # Linter
mypy>=1.5.0                   # Type checking

# ========================
# UTILITÁRIOS
# ========================
python-dotenv>=1.0.0          # Variáveis de ambiente (.env)
colorama>=0.4.6               # Cores no terminal (Windows)

# ========================
# PRODUÇÃO (Opcional)
# ========================
# gunicorn>=21.0.0            # WSGI server para produção
# python-socketio>=5.9.0      # WebSocket com fallback
# redis>=5.0.0                # Cache/Session storage

# ========================
# DOCUMENTAÇÃO (Opcional)
# ========================
# sphinx>=7.0.0               # Geração de documentação
# sphinx-rtd-theme>=1.3.0     # Tema ReadTheDocs

# ========================
# NOTAS DE VERSÃO
# ========================
# - Python 3.9+ é obrigatório
# - Testado em Windows 10/11
# - Todas as dependências são open-source
# - Tamanho total: ~200 MB (incluindo venv)
